requirements.txt
boto3
pandas
mlflow
scikit-learn
evidently
streamlit
pytest
jupyter
python-dotenv
yfinance
pyarrow
s3fs
plotly
dagshub
# pandas-ta
fastparquet
fastapi
uvicorn
requests
gradio
